Changes made to add in a FindMethod and fixed MobileNumber

+++ CarSellersClasses/clsStaff.cs
@@ -29,8 +29,8 @@ namespace CarSellersClasses
             set { mSurname = value; }
         }
 
-        private int mMobileNumber;
-        public int MobileNumber
+        private Int64 mMobileNumber;
+        public Int64 MobileNumber
         {
             get { return mMobileNumber; }
             set { mMobileNumber = value; }
@@ -86,7 +86,7 @@ namespace CarSellersClasses
                 mFirstName = Convert.ToString(Database.DataTable.Rows[0]["FirstName"]);
                 mSurname = Convert.ToString(Database.DataTable.Rows[0]["Surname"]);
                 mDOB = DateTime.Parse(Convert.ToString(Database.DataTable.Rows[0]["DateOfBirth"]));
-                mMobileNumber = Convert.ToInt32(Database.DataTable.Rows[0]["MobileNumber"]);
+                mMobileNumber = Convert.ToInt64(Database.DataTable.Rows[0]["MobileNumber"]);
                 mInOffice = Convert.ToBoolean(Database.DataTable.Rows[0]["InOffice"]);
                 mDelete = Convert.ToBoolean(Database.DataTable.Rows[0]["toDelete"]);
                 return true;
